Crate bucket

The bucket crate is just a playground for Rust experiments.

Initially we're going to use the csv crate to read from a .csv file. We'll use csv/test.csv for now.

We're also going to experiment with file organization and module structure, and with documenting the code.

name,population,latitude,longitude
New York ,8287238,40.7305991,-73.9865812
Los Angeles ,3826423,34.053717,-118.2427266
...

Modules

For now have broken out the following structure:

src/
├── datasets
│   └── mod.rs
├── import
│   ├── mod.rs
│   └── reader.rs
└── main.rs

Modules

datasets
The datasets module is for describing data structures.

import
The imports module is for collecting import functionality.

Functions

get_config
Get a param's value from the Settings.toml config file.

read_csv
Read data from the CSV files.

read_database
Read data from the Postgres database.